backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
pydantic[email]==2.12.5
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pyotp==2.9.0
requests==2.32.3
